requirements-dev.txt
# Testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Parallel execution and timeout protection
pytest-xdist==3.5.0
pytest-timeout==2.2.0

# Coverage reporting
coverage[toml]==7.3.3

# HTTP mocking for some tests
responses==0.24.1
